VirtualPlayer and ShadowPlayer models created; Balance, Production, Player, Resources, Building, and Technology models updated accordingly
