+++ back/game/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from .models import SourceBuilding, SourceEvent, SourceTechnology
 from .models import Game, Event, HydrocarbonSupplyPile
 from .models import Profile
-from .models import Player, Balance, Building, Production, Resources, Technology
+from .models import VirtualPlayer, Player, ShadowPlayer, Balance, Building, Production, Resources, Technology
 
 admin.site.register(SourceBuilding)
 admin.site.register(SourceEvent)
@@ -17,4 +17,6 @@ admin.site.register(Building)
 admin.site.register(Production)
 admin.site.register(Resources)
 admin.site.register(Technology)
+# admin.site.register(VirtualPlayer)
+admin.site.register(ShadowPlayer)
 # To be continued...
