+++ back/game/models/__init__.py
@@ -5,7 +5,10 @@ from .source_building import SourceBuilding
 from .game import Game
 from .event import Event
 from .hydrocarbon_supply_pile import HydrocarbonSupplyPile
+from .virtual_player import VirtualPlayer
 from .player import Player
+from .shadow_player import ShadowPlayer
+
 
 from .profile import Profile
 from .balance import Balance
