+++ back/game/models/player.py
@@ -10,20 +10,23 @@ from .technology import Technology
 from .source_building import SourceBuilding
 from .source_technology import SourceTechnology
 
+from .virtual_player import VirtualPlayer
 
-class Player(models.Model):
+
+class Player(VirtualPlayer):
     """
-    Player model
+    Player model, inherit from VirtualPlayer model
 
-    Fields :
-        game (Game) : ForeignKey link to the game in which the player plays
-        profile (Profile) : profile which controls the player
+    Inherited fields :
+        * balance (OneToOneField <- Balance): player balance
+        * resources (OneToOneField <- Resources): resources owned by the player
+        * production (OneToOneField <- Production): player production at the beginning of each generation
+        * technologies (ForeignKey <- PlayerTechnology): technologies
+        * buildings (ForeignKey <- PlayerBuilding): buildings
 
-        balance (OneToOneField <- Balance): player balance
-        resources (OneToOneField <- Resources): resources owned by the player
-        production (OneToOneField <- Production): player production at the beginning of each generation
-        technologies (ForeignKey <- PlayerTechnology): technologies
-        buildings (ForeignKey <- PlayerBuilding): buildings
+    New fields :
+        * game (Game) : ForeignKey link to the game in which the player plays
+        * profile (Profile) : profile which controls the player
     """
     game = models.ForeignKey('Game', on_delete=models.CASCADE, related_name="players", editable=False)
     profile = models.ForeignKey('Profile', on_delete=models.CASCADE, related_name="players", editable=False)
@@ -34,19 +37,6 @@ class Player(models.Model):
     def __eq__(self, other):
         return (self.game.id == other.game.id and
                 self.profile.id == other.profile.id)
-
-    def has_same_possessions(self, other):
-        if (self.balance != other.balance or
-           self.production != other.production or
-           self.resources != other.resources):
-            return False
-        for building in self.buildings.all():
-            if building != Building.objects.get(index=building.index, player=other):
-                return False
-        for technology in self.technologies.all():
-            if technology != Technology.objects.get(index=technology.index, player=other):
-                return False
-        return True
 
     @classmethod
     def create(cls, profile, game):
@@ -100,23 +90,3 @@ class Player(models.Model):
     def green_income(self):
         """ Apply the environment generation income to the environment balance """
         self.balance.green_income()
-
-    def purchase_building(self, slug):
-        """ Purchase the building with given slug if possible. If not, return an error string. """
-        building = self.buildings.get(slug=slug)
-        (is_purchasable, error_message) = building.is_purchasable()
-        if is_purchasable:
-            building.copies += 1
-            building.trigger_post_purchase_effects()
-            building.save()
-        return (building, error_message)
-
-    def purchase_technology(self, slug):
-        """ Purchase the technology with given slug if possible. If not, return an error string. """
-        technology = self.technologies.get(slug=slug)
-        (is_purchasable, error_message) = technology.is_purchasable()
-        if is_purchasable:
-            technology.purchased = True
-            technology.trigger_post_purchase_effects()
-            technology.save()
-        return (technology, error_message)
