+++ back/game/models/virtual_player.py
@@ -1,0 +1,57 @@
+from django.db import models
+
+from .. import game_settings as constant
+
+from .resources import Resources
+from .production import Production
+from .balance import Balance
+from .building import Building
+from .technology import Technology
+from .source_building import SourceBuilding
+from .source_technology import SourceTechnology
+
+
+class VirtualPlayer(models.Model):
+    """
+    Virtual Player model; father model of ShadowPlayer and Player
+
+    Fields :
+        * balance (OneToOneField <- Balance): player balance
+        * resources (OneToOneField <- Resources): resources owned by the player
+        * production (OneToOneField <- Production): player production at the beginning of each generation
+        * technologies (ForeignKey <- PlayerTechnology): technologies
+        * buildings (ForeignKey <- PlayerBuilding): buildings
+    """
+
+    def has_same_possessions(self, other):
+        if (self.balance != other.balance or
+           self.production != other.production or
+           self.resources != other.resources):
+            return False
+        for building in self.buildings.all():
+            if building != Building.objects.get(index=building.index, player=other):
+                return False
+        for technology in self.technologies.all():
+            if technology != Technology.objects.get(index=technology.index, player=other):
+                return False
+        return True
+
+    def purchase_building(self, slug):
+        """ Purchase the building with given slug if possible. If not, return an error string. """
+        building = self.buildings.get(slug=slug)
+        (is_purchasable, error_message) = building.is_purchasable()
+        if is_purchasable:
+            building.copies += 1
+            building.trigger_post_purchase_effects()
+            building.save()
+        return (building, error_message)
+
+    def purchase_technology(self, slug):
+        """ Purchase the technology with given slug if possible. If not, return an error string. """
+        technology = self.technologies.get(slug=slug)
+        (is_purchasable, error_message) = technology.is_purchasable()
+        if is_purchasable:
+            technology.purchased = True
+            technology.trigger_post_purchase_effects()
+            technology.save()
+        return (technology, error_message)
